requirements.txt
numpy
pandas
scipy
matplotlib
tkinter
reportlab
scikit-learn
pyserial
